requirements.txt
django
psycopg2
djangorestframework==3.14.0  # o la versión más reciente
python-decouple
django-cors-headers==4.7.0